pyproject.toml
[project]
name = "telraam-viz"
version = "0.1.0"
description = "Compare Telraam traffic data across custom time periods"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30",
    "pandas>=2.0",
    "plotly>=5.18",
    "requests>=2.31",
    "pyarrow>=14.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
